Rename generate-parser.js to parser.js
