+++ index.js
@@ -1,4 +1,4 @@
-var parser = require('./generate-parser')
+var parser = require('./parser')
 
 module.exports = function (argument) {
   return parser.parse(argument)
